refactor: Server into module
